+++ src/Components/Navbar/Navbar.js
@@ -5,6 +5,7 @@ import styles from "../Navbar/Navbar.module.scss";
 
 // ? All Component and View imports
 import LogIn from "../../Components/LogIn/LogIn";
+import SignUp from "../../Components/SignUp/Signup";
 // import LogIn2 from "../test/LogIn2/LogIn2";
 
 const Navbar = () => {
@@ -21,6 +22,34 @@ const Navbar = () => {
         console.log("background was clicked close log in");
     };
 
+    const [showSignUp, setShowSignUp] = useState(false);
+
+    const openSignUp = (event) => {
+        event.preventDefault();
+        setShowSignUp(true);
+    };
+
+    const closeSignUp = (event) => {
+        event.preventDefault();
+        setShowSignUp(false);
+        console.log("background was clicked close sign up");
+    };
+
+    // Login Modal is not visible
+    if (!props.showLogin) {
+        return null;
+    }
+
+    const stop = (event) => {
+        event.stopPropagation();
+    };
+
+    // !! ASK GROUP, go to line
+    const switchToSignup = (event) => {
+        props.setShowLogIn(false);
+        setShowSignUp(true);
+    };
+
     return (
         <>
             <div className={styles["navbar-container"]}>
@@ -29,7 +58,7 @@ const Navbar = () => {
                     showLogin={showLogIn}
                     // setShowLogIn={setShowLogIn}
                 />
-
+                <SignUp onCloseSignup={closeSignUp} showSignup={showSignUp} />
                 <div className={styles["icon-container"]}>
                     <div className={styles.icon}>
                         <i class="fas fa-caravan"></i>
